Core: move block import logic to BlockchainImpl
